microservices/input_integration_pipeline/requirements.txt
requests
fastapi
pytest 
azure-storage-blob
azure-core
pydantic
python-dotenv
uvicorn
pandas
aiohttp
pyarrow
fastparquet
tenacity
httpx